refactor: 接口兼容
调整SetSelfProfile接口为SetQQProfile，使其兼容gocq标准
